backend_django/requirements.txt
asgiref==3.5.2
Django==3.2.14
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.1
numpy==1.23.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
